Gracefully close with hotkey when Scratch has been modified
With autosave disabled, if a user modifies their scratch and uses the 
hotkey to close it, the TextEditor is destroyed before the user is 
prompted to save. When the user then clicks save in the prompt, they see 
an error:

> 'scratch' could not be saved.
> Error: undefined

This change uses Pane's public `destroyItem` function to ensure that 
when a user without autosaved enabled closes their scratch with the 
hotkey, it will remain open until they have chosen "Save" or "Don't 
Save," in which case the scratch will be closed, or "Cancel," in which 
case the scratch will remain open. This is the same behavior as when the 
user clicks the "X" on the tab to close their scratch. It does not 
impact the behavior for users with autosave enabled.

closes https://github.com/destradax/scratch/issues/13